+++ TLSharp.Core/Exceptions/CloudPasswordNeededException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TLSharp.Core.Exceptions
+{
+    public class CloudPasswordNeededException : Exception
+    {
+        internal CloudPasswordNeededException(string msg) : base(msg) { }
+    }
+}
