+++ TLSharp.Core/Exceptions/InvalidPhoneCodeException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TLSharp.Core.Exceptions
+{
+    public class InvalidPhoneCodeException : Exception
+    {
+        internal InvalidPhoneCodeException(string msg) : base(msg) { }
+    }
+}
