+++ TLSharp.Core/Exceptions/MissingApiConfigurationException.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace TLSharp.Core.Exceptions
+{
+    public class MissingApiConfigurationException : Exception
+    {
+        public const string InfoUrl = "https://github.com/sochix/TLSharp#quick-configuration";
+
+        internal MissingApiConfigurationException(string invalidParamName) :
+            base($"Your {invalidParamName} setting is missing. Adjust the configuration first, see {InfoUrl}")
+        {
+        }
+    }
+}
