+++ TLSharp.Core/Network/Exceptions/UserMigrationException.cs
@@ -1,0 +1,10 @@
+﻿namespace TLSharp.Core.Network.Exceptions
+{
+    internal class UserMigrationException : DataCenterMigrationException
+    {
+        internal UserMigrationException(int dc)
+            : base($"User located on a different DC: {dc}.", dc)
+        {
+        }
+    }
+}
