+++ TLSharp.Core/TelegramClient.cs
@@ -12,8 +12,10 @@ using TeleSharp.TL.Help;
 using TeleSharp.TL.Messages;
 using TeleSharp.TL.Upload;
 using TLSharp.Core.Auth;
+using TLSharp.Core.Exceptions;
 using TLSharp.Core.MTProto.Crypto;
 using TLSharp.Core.Network;
+using TLSharp.Core.Network.Exceptions;
 using TLSharp.Core.Utils;
 using TLAuthorization = TeleSharp.TL.Auth.TLAuthorization;
 
@@ -445,24 +447,5 @@ namespace TLSharp.Core
                 _transport = null;
             }
         }
-    }
-
-    public class MissingApiConfigurationException : Exception
-    {
-        public const string InfoUrl = "https://github.com/sochix/TLSharp#quick-configuration";
-
-        internal MissingApiConfigurationException(string invalidParamName) :
-            base($"Your {invalidParamName} setting is missing. Adjust the configuration first, see {InfoUrl}")
-        {
-        }
-    }
-
-    public class InvalidPhoneCodeException : Exception
-    {
-        internal InvalidPhoneCodeException(string msg) : base(msg) { }
-    }
-    public class CloudPasswordNeededException : Exception
-    {
-        internal CloudPasswordNeededException(string msg) : base(msg) { }
     }
 }
